requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Scientific computing
numpy>=1.24.0

# Optional: For GPU support (uncomment if you have CUDA)
# torch==2.0.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# torchvision==0.15.0+cu118 --index-url https://download.pytorch.org/whl/cu118

# Optional: Accelerate for faster model loading
accelerate>=0.20.0